Update score system

+++ Assets/Script/UIManager.cs
@@ -18,17 +18,17 @@ public class UIManager : MonoBehaviour
         _instance = this;
     }
 
-    public void SetScore(float score)
+    public void SetScore(int score)
     {
         _scoreText.text = $"Distance parcourue: {score}";
     }
 
-    public void SetBestScore(float score)
+    public void SetBestScore(int score)
     {
         _bestScoreText.text = $" Plus loin attein: {score}";
     }
 
-    public void SetLastScore(float score)
+    public void SetLastScore(int score)
     {
         _lastScoreText.text = $"Derniere distance atteinte: {score}";
     }
